requirements.txt
f5-sdk
logbook
PyYaml==5.4.1